# Small speedup for matrix multiplication