From [this SO answer](https://stackoverflow.com/questions/35923787/fast-large-matrix-multiplication-in-r):
``` r
n_cols = 30

my_J = matrix(sample(1:50, n_cols * 2500000, TRUE), ncol = n_cols)
my_V = matrix(sample(1:10, n_cols^2, TRUE), ncol = n_cols)

bench::mark(
  my_J %*% my_V,
  crossprod(t(my_J), my_V),
  iterations = 10
)
#> # A tibble: 2 × 6
#>   expression                    min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>               <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 my_J %*% my_V               3.35s    3.35s     0.299    1.12GB     2.99
#> 2 crossprod(t(my_J), my_V)    2.08s    2.08s     0.481     1.4GB     5.29
```

Doesn't really make sense to benchmark with `marginaleffects` functions since it's a small change in the total time

---

Besides this, I'm seeing `data.table` warnings in the tests (even before this change) but I don't know where they come from.
```
1: In data.table::setDT(modeldata) :
  Some columns are a multi-column type (such as a matrix column): [3]. setDT will retain 
these columns as-is but subsequent operations like grouping and joining may fail. Please
 consider as.data.table() instead which will create a new column for each embedded column.
```