@vincentarelbundock 2023-03-06
This is interesting. I wonder what the priority should be here: speed or memory. See last line which is a bit more readable and more memory-efficient:

``` r
n_cols = 30
J = matrix(rnorm(n_cols * 2500000), ncol = n_cols)
V = matrix(runif(n_cols^2), ncol = n_cols)

bench::mark(
    colSums(t(J %*% V) * t(J)),
    colSums(t(crossprod(t(J), V)) * t(J)),
    rowSums(tcrossprod(J, V) * J),
    iterations = 5
)

# # A tibble: 3 × 6
#   expression                                 min   median `itr/sec` mem_alloc `gc/sec`
#   <bch:expr>                            <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
# 1 colSums(t(J %*% V) * t(J))               2.88s    3.17s     0.313     1.7GB    0.313
# 2 colSums(t(crossprod(t(J), V)) * t(J))    2.68s     3.1s     0.326    2.25GB    0.521
# 3 rowSums(tcrossprod(J, V) * J)            2.99s    3.34s     0.302    1.14GB    0.181
```

The warning is probably in the `mlogit` or `mblogit` tests. Could diagnose by setting `options(warn=2)` before testing. My guess is that this is not an important warning and we chould just suppress the warning around that `setDT` call.