@vincentarelbundock 2023-03-06
I'm asking because this operation is only executed once, and the difference is a fraction of a second for massive datasets.