@etiennebacher 2023-03-06
Interesting, you have very small time differences while in my case the gap is around 30%. But in any case this is indeed a small change and it consumes more memory. Maybe it's not worth changing it.

FYI, I think I'll stop the optimization here because it seems 90-95% of the time spent is now in `get_predict()` in all `marginaleffects` functions (except maybe for bayesian models, I didn't look at them at all).

Just to provide a complete benchmark if someone finds it useful:
```r
bench::press(
  n_cols = c(1, 10, 20, 30, 40),
  {
    J = create_J(n_cols)
    V = create_V(n_cols)
    bench::mark(
      colSums(t(J %*% V) * t(J)),
      colSums(t(crossprod(t(J), V)) * t(J)),
      rowSums(tcrossprod(J, V) * J),
      iterations = 5
    )
  }
)

Running with:
  n_cols
1      1
2     10
3     20
4     30
5     40
# A tibble: 15 × 14
   expression                            n_cols      min   median `itr/sec` mem_al…¹ gc/se…² n_itr  n_gc total_…³ result memory     time      
   <bch:expr>                             <dbl> <bch:tm> <bch:tm>     <dbl> <bch:by>   <dbl> <int> <dbl> <bch:tm> <list> <list>     <list>    
 1 colSums(t(J %*% V) * t(J))                 1  30.41ms  34.97ms    28.8    76.29MB   0         5     0 173.42ms <dbl>  <Rprofmem> <bench_tm>
 2 colSums(t(crossprod(t(J), V)) * t(J))      1  32.41ms  33.92ms    26.1    95.37MB   0         5     0 191.67ms <dbl>  <Rprofmem> <bench_tm>
 3 rowSums(tcrossprod(J, V) * J)              1   34.8ms  40.27ms    24.8    57.22MB   0         5     0 201.28ms <dbl>  <Rprofmem> <bench_tm>
 4 colSums(t(J %*% V) * t(J))                10 512.02ms 530.88ms     1.82  591.28MB   0.456     4     1    2.19s <dbl>  <Rprofmem> <bench_tm>
 5 colSums(t(crossprod(t(J), V)) * t(J))     10 472.36ms 472.36ms     2.12  782.01MB   8.47      1     4 472.36ms <dbl>  <Rprofmem> <bench_tm>
 6 rowSums(tcrossprod(J, V) * J)             10 480.84ms 521.93ms     1.93  400.54MB   0.484     4     1    2.07s <dbl>  <Rprofmem> <bench_tm>
 7 colSums(t(J %*% V) * t(J))                20    1.77s    1.83s     0.553   1.14GB   0.553     5     5    9.04s <dbl>  <Rprofmem> <bench_tm>
 8 colSums(t(crossprod(t(J), V)) * t(J))     20    1.28s    1.43s     0.711   1.51GB   0.996     5     7    7.03s <dbl>  <Rprofmem> <bench_tm>
 9 rowSums(tcrossprod(J, V) * J)             20    1.56s    1.61s     0.596 782.01MB   0.119     5     1    8.39s <dbl>  <Rprofmem> <bench_tm>
10 colSums(t(J %*% V) * t(J))                30    3.57s    3.63s     0.271    1.7GB   0.379     5     7   18.47s <dbl>  <Rprofmem> <bench_tm>
11 colSums(t(crossprod(t(J), V)) * t(J))     30    2.61s    2.72s     0.369   2.25GB   0.369     5     5   13.55s <dbl>  <Rprofmem> <bench_tm>
12 rowSums(tcrossprod(J, V) * J)             30    3.29s    3.59s     0.284   1.14GB   0.171     5     3   17.58s <dbl>  <Rprofmem> <bench_tm>
13 colSums(t(J %*% V) * t(J))                40    5.98s    6.12s     0.164   2.25GB   0.262     5     8   30.56s <dbl>  <Rprofmem> <bench_tm>
14 colSums(t(crossprod(t(J), V)) * t(J))     40    4.26s    4.43s     0.225      3GB   0.496     5    11    22.2s <dbl>  <Rprofmem> <bench_tm>
15 rowSums(tcrossprod(J, V) * J)             40    5.66s    5.81s     0.172   1.51GB   0.138     5     4   29.03s <dbl>  <Rprofmem> <bench_tm>
```