@vincentarelbundock 2023-03-06
Thanks for those. The variation is quite interesting.

> FYI, I think I'll stop the optimization here because it seems 90-95% of the time spent is now in get_predict() in all marginaleffects functions

Yeah, that makes sense. Thanks so much for all of those. I think in the last two versions we probably got speed-ups of 6-8x for simple GLM models. (Probably a bit less in other packages, because their `predict` methods tend to be less optimized, and we spend even more time in `get_predict()`). So this is very very cool.

For `lm()` models we could probably get a 2x speedup by building our own design matrix and multiplying it, but that would probably be a bit less safe. Overall, rolling out our own faster `predict()` methods feels like a bit of a rabbit hole (and also dangerous), so perhaps that is the end of optimization for now.